refactor: lottoNumbersBox와 관련된 파일을 분류하는 디렉터리 생성
