+++ src/view/web/layers/lottoNumbersBox/lottoNumbersBox.css
